==> 1-series.ipynb <==
import pandas as pd

dados = [10,20,30,40,50]
series = pd.Series(dados)
series

#vendas mensais de uma loja 
vendas_mensais = pd.Series(
    [200,2500,3000,2800,3500,4000,4500,5000,4800,5200,6000,7000],
    index=["jan",'fev','mar','abr','mai','jun','jul','ago','set','out','nov','dez']
)

vendas_mensais

#calcular e exibir a média de vendas 
media_vendas = vendas_mensais.mean()
media_vendas

#identificando mes com maor venda 
mes_maior_venda = vendas_mensais.idxmax()
mes_maior_venda

#exibir vendas acima da média 
vendas_acima_media = vendas_mensais[vendas_mensais > media_vendas]
vendas_acima_media
