==> 2-dataframe.ipynb <==
import pandas as pd

dados = {
    "Nome": ['alice','bob','charlie'],
    "Idade":[24,30,22],
    "Cidade":['São Paulo','Rio de Janeiro','Belo Horizonte']
}
dados

df = pd.DataFrame(dados)
df

df.head(1)

df.tail(1)

dados_notas = {
    'nome':['alice','bob','charlie','david','eva'],
    'matematica':[58,90,78,88,92],
    'fisica':[80,75,88,85,95],
    'quimica':[78,82,85,80,88]
}

df_notas = pd.DataFrame(dados_notas)
df_notas

#calcular a méddia das notas por aluno 
df_notas['Média'] = df_notas[["matematica","fisica","fisica"]].mean(axis=1)
df_notas

#identificar alunos aprovados com nota excelente ( acuma de 83)
alunos_aprovados = df_notas[df_notas["Média"] >= 83]
alunos_aprovados

#Exibir a média e os alunos aprovados 
df_notas[["nome",'Média']]


alunos_aprovados[['nome', 'Média']]
