==> 5-importa_csv.ipynb <==
import pandas as pd

df_produtos = pd.read_csv("dados/Produtos.csv" , sep=',')#por padrão é "," 
df_produtos

df_produtos.head(8)

df_produtos.tail(9)

df_produtos.shape

df_produtos.dtypes

df_produtos.sort_values(by='Quantidade', ascending=False)

df_produtos['Categoria'].value_counts()
