==> 6-analize_csv.ipynb <==
import pandas as pd

df_produtos = pd.read_csv("dados/Produtos.csv" ) 
df_produtos

df_produtos.dtypes

df_produtos['Qantidade'] = pd.to_numeric(df_produtos['Quantidade'])
df_produtos['Preço'] = pd.to_numeric(df_produtos['Preço'])



vendas_categoria = df_produtos.groupby('Categoria')['Quantidade'].sum()
vendas_categoria

#produto mais vendido 
produto_mais_vendido = df_produtos.groupby('Produto')['Quantidade'].sum().idxmax()
produto_mais_vendido

produto_mais_vendido = df_produtos.groupby('Produto')['Quantidade'].sum()
produto_mais_vendido

media_preco_produto = df_produtos.groupby('Produto')['Preço'].mean()
media_preco_produto

#Correlação entre Unidades Vendidas e preço Unitário
correlacao = df_produtos['Quantidade'].corr(df_produtos["Preço"])
correlacao

vendas_categoria_df = vendas_categoria.reset_index()
vendas_categoria_df.columns = ['Categoria', 'UnidadesVendidas']
vendas_categoria_df

vendas_categoria_df.to_csv("dados/vendas_categoria.csv", index=False)
