==> 8-analize_planilha.ipynb <==
import pandas as pd

df = pd.read_excel('dados/tabela_classificacao.xlsx')

df.dtypes

df.describe()

df.head(2)

df.tail(2)

# Analize de desempenho por time 
desempenho_info = df[['Nome','pontos','vitorias','saldo_gols','gols_pro','gols_contra']].set_index('Nome')
desempenho_info



time_com_mais_vitorias = df[df['vitorias'] == df['vitorias'].max()]
time_com_mais_vitorias

df['gols_pro'].value_counts()

correlacao = df['gols_pro'].corr(df['pontos'])
correlacao
